geom_waffle {waffle}	R Documentation

Waffle (Square pie chart) Geom

Description

There are two special/critical aes() mappings:

fill (so the geom knows which column to map the fills to)

values (which column you're mapping the filling for the squares with)

Usage

geom_waffle(mapping = NULL, data = NULL, n_rows = 10,
  make_proportional = FALSE, flip = FALSE, na.rm = NA,
  show.legend = NA, radius = grid::unit(0, "npc"),
  inherit.aes = TRUE, ...)

GeomWaffle

stat_waffle(mapping = NULL, data = NULL, geom = "waffle",
  n_rows = 10, make_proportional = FALSE, flip = FALSE,
  radius = grid::unit(0, "npc"), na.rm = NA, show.legend = NA,
  inherit.aes = TRUE, ...)

StatWaffle

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame., and will be used as the layer data.

n_rows
how many rows should there be in the waffle chart? default is 10

make_proportional
compute proportions from the raw values? (i.e. each value n will be replaced with n/sum(n)); default is FALSE.

flip
If TRUE, flip x and y coords. n_rows then becomes n_cols. Useful to achieve waffle column chart effect. Defaults is FALSE.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

radius
radius for round squares

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

...
other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.

geom
geom to use (default is "waffle")

Format

An object of class GeomWaffle (inherits from GeomRtile, GeomRrect, Geom, ggproto, gg) of length 5.

Examples

data.frame(
  parts = factor(rep(month.abb[1:3], 3), levels=month.abb[1:3]),
  vals = c(10, 20, 30, 6, 14, 40, 30, 20, 10),
  fct = c(rep("Thing 1", 3), rep("Thing 2", 3), rep("Thing 3", 3))
) -> xdf

ggplot(xdf, aes(fill = parts, values = vals)) +
  geom_waffle() +
  facet_wrap(~fct)

[Package waffle version 1.0.1 Index]